% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot accident locations on a US map using data from the Fatality Analysis Report System.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The state identification number used by the Fatality Analysis Report System.}

\item{year}{The year for which the data are to be plotted.}
}
\value{
A state map with points indicate accident locations for the specified year.
}
\description{
Plot accident locations on a US map using data from the Fatality Analysis Report System.
}
\details{
The function will search and retrive the requested data from the Fatality Analysis Report System.  An error will be returned if there state number is invalid.  If there is no accident data to report for the specified year, this will be reported without a map.
}
\examples{
\dontrun{
 fars_map_state(42,2013)}
}
