% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read in a .csv file and format as a tibble dataframe}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{The path name for the file you wish to read in.}
}
\value{
A tibble dataframe of the read in data.
}
\description{
Read in a .csv file and format as a tibble dataframe
}
\note{
Will return a message if the file cannot be found.
}
