% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read in the data from the Fatality Analysis Reporting System for the years specified.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector of years}
}
\value{
Mutiple tibble dataframes of data for the requested years.
}
\description{
Read in the data from the Fatality Analysis Reporting System for the years specified.
}
\details{
For each year, the fears_read_years function will search for a corresponding file with the filename formatted as "accident_date.csv.bz2", insterting the year for date. If the file does not exist, it will return an error indicate the year is invalid.
}
\examples{
\dontrun{
these_years <- c(2013,2014,2015)
fars_read_years(these_years)}
}
