% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Create a summary table of the number of accidents per year, according to the Fatality #' Analysis Report System for the years specified.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector of years}
}
\value{
A table of counts of accidents for each year requested by month.
}
\description{
Create a summary table of the number of accidents per year, according to the Fatality #' Analysis Report System for the years specified.
}
\details{
The function will read in all data files for the years requested and produce a summary table for the number of accidents that occurred each month for the years specified.
}
\examples{
\dontrun{
these_years <- c(2013,2014,2015)
fars_summarize_years(these_years)}
}
