% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename and set the year to be used in the name "accident_date.csv.bz2".}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year to be referenced in the file name}
}
\value{
A character string of the filename.
}
\description{
Create a filename and set the year to be used in the name "accident_date.csv.bz2".
}
